+++ Cargo.toml
@@ -5,4 +5,5 @@ members = [
   "rust-web-handson-domain",
   "rust-web-handson-infra",
   "rust-web-handson-util",
+  "integration-test",
 ]
