+++ integration-test/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "integration-test"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+reqwest = { version = "0.11", features = ["json", "blocking"] }
+tokio = {version = "1.18.2", features = ["full"]}
