+++ rust-web-handson-presentation/Cargo.toml
@@ -17,7 +17,6 @@ thiserror = "1.0.31"
 tokio = {version = "1.18.2", features = ["full"]}
 tracing = "0.1.34"
 tracing-subscriber = {version = "0.3.11", features = ["env-filter"]}
-reqwest = { version = "0.11", features = ["json", "blocking"] }
 
 [dev-dependencies]
 mockall = "0.11.1"
