+++ sonar.properties
@@ -1,0 +1,1 @@
+community.rust.clippy.reportPaths="./clippy-report-file"
